Added Chapter 4 - Loops, Arrays, Strings, Functions, and Practice Programs

+++ CHAPTER 1/01_variables.c
@@ -8,8 +8,8 @@
 
 #include<stdio.h>
 int main() {
-    int a;  // Variable declaration
+    float a;  // Variable declaration
     a = 6;  // Variable initialization
-    printf("The output of this program is %d", a);   
+    printf("The output of this program is %f", a);   
     return 0; 
 }
